% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_stoc_args}
\alias{check_stoc_args}
\title{#'@export
check_SCSD_args <- function(ARGS, N, D){}
\usage{
check_stoc_args(ARGS, N, D)
}
\description{
\if{html}{\out{<div class="sourceCode">}}\preformatted{out <- ARGS
}\if{html}{\out{</div>}}
}
\details{

}
