% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_tidy_path3}
\alias{get_tidy_path3}
\title{Get parameters trajectories}
\usage{
get_tidy_path3(MOD_OBJ, PATH_LAB)
}
\arguments{
\item{MOD_OBJ}{object from \link{fit_isingGraph3}}

\item{PATH_LAB}{Label of the path of interest}
}
\description{
Get parameters trajectories
}
