% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ising_from_graph_to_theta}
\alias{ising_from_graph_to_theta}
\title{From matrix to vector}
\usage{
ising_from_graph_to_theta(graph, intercepts)
}
\arguments{
\item{graph}{Graph matrix}

\item{intercepts}{Nodes intercepts}
}
\description{
From matrix to vector
}
