% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ising_from_theta_to_emat}
\alias{ising_from_theta_to_emat}
\title{From vector to matrix}
\usage{
ising_from_theta_to_emat(par, p)
}
\arguments{
\item{par}{Parameter vector}

\item{p}{Number of nodes}
}
\description{
From vector to matrix
}
