% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ncl}
\alias{ncl}
\title{Negative composite log-likelihood}
\usage{
ncl(DATA, THETA, CONSTRAINTS, VERBOSEFLAG = FALSE)
}
\arguments{
\item{DATA}{Data matrix with \code{n} rows and \code{p} columns}

\item{THETA}{Parameter vector.}

\item{CONSTRAINTS}{Vector of booleans: TRUE denotes free-to-estimate parameters}

\item{VERBOSEFLAG}{Verbose output.}
}
\description{
Compute the negative composite log-likelihood of an Ising graph and its gradient
}
