% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sampleJ}
\alias{sampleJ}
\title{Compute sample variability matrix}
\usage{
sampleJ(THETA, DATA, CONSTRAINTS, VERBOSEFLAG = FALSE)
}
\arguments{
\item{THETA}{Parameter vector}

\item{DATA}{Matrix with \code{n} rows and \code{p} columns}

\item{CONSTRAINTS}{Vector of booleans: TRUE denotes free-to-estimate parameters}

\item{VERBOSEFLAG}{Verbose output}
}
\description{
Compute sample variability matrix
}
