% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sampleVar}
\alias{sampleVar}
\title{Compute sample covariance matrix}
\usage{
sampleVar(THETA, DATA, CONSTRAINTS, NU, METHOD, RANGE, TOTFLAG, PRINTFLAG)
}
\arguments{
\item{THETA}{Parameter vector}

\item{DATA}{Matrix with \code{n} rows and \code{p} columns}

\item{CONSTRAINTS}{Vector of booleans: TRUE denotes free-to-estimate parameters}

\item{NU}{Number of pairs per iteration on average.}

\item{METHOD}{Allowed choices are "ucminf", "standard", "bernoulli", "hyper",
"recycle_standard", "recycle_bernoulli", "recycle_hyper"}

\item{RANGE}{Number of iterations after the burn-in.}

\item{TOTFLAG}{Compute total variance}

\item{PRINTFLAG}{Verbose output}
}
\description{
Compute sample covariance matrix
}
